Adds build step to Dockerfile
The old Dockerfile relied on someone to build the app before
running docker build. Now, the Dockerfile runs all necessary steps.

Stage 0 is a "build-stage", based on Node.js, to build and compile
the application.
Stage 1, based on Nginx, to have only the compiled app, ready for
production with Nginx.